beyondweb/requirements.txt
asgiref==3.11.0
diff-match-patch==20241021
dj-database-url==3.0.1
Django==5.2.9
django-import-export==4.3.14
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
tablib==3.9.0
whitenoise==6.6.0
